+++ 2017-2018/ROS_RoboBuggy/src/robobuggy/include/imu/IMU_Broadcaster.h
@@ -6,11 +6,12 @@ class IMU_Broadcaster
 {
 public:
     IMU_Broadcaster();
-    ros::NodeHandle nh;
-    ros::Publisher imu_pub;
     float get_spoofed_x_accel();
     float get_spoofed_y_accel();
     float get_spoofed_z_accel();
     void publish_new_spoofed_message();
+private:
+    ros::NodeHandle nh;
+    ros::Publisher imu_pub;
 }
 ;
