+++ 2017-2018/ROS_RoboBuggy/src/robobuggy/src/encoder/ENC_Broadcaster.cpp
@@ -2,40 +2,25 @@
 // Created by bhai on 8/17/17.
 //
 
-#include <ros/init.h>
-#include <ros/node_handle.h>
-#include <robobuggy/ENC.h>
+#include "encoder/ENC_Broadcaster.h"
 
-int main(int argc, char **argv)
+ENC_Broadcaster::ENC_Broadcaster()
 {
-    ros::init(argc, argv, "ENC_Broadcaster");
-
-    ros::NodeHandle nh;
-
-    ros::Publisher enc_pub = nh.advertise<robobuggy::ENC>("ENC", 1000);
-
-    ros::Rate loop_rate(10);
-
-    float distance = 0.0;
-
-    // infinitely spin sending dummy encoder messages
-    while (ros::ok())
-    {
-        robobuggy::ENC msg;
-
-        msg.distance_m = distance;
-        msg.velocity_ms = 1;
-        msg.acceleration_mss = 0;
-        msg.raw_data_word = 1337;
-
-        distance += 0.1;
-
-        enc_pub.publish(msg);
-
-        ros::spinOnce();
-
-        loop_rate.sleep();
-    }
-
-    return 0;
+    enc_pub = nh.advertise<robobuggy::ENC>("ENC", 1000);
 }
+
+void ENC_Broadcaster::publish_spoof_message()
+{
+    robobuggy::ENC msg;
+
+    msg.distance_m = spoof_distance;
+    msg.velocity_ms = 1;
+    msg.acceleration_mss = 0;
+    msg.raw_data_word = 1337;
+
+    spoof_distance += 0.1;
+
+    enc_pub.publish(msg);
+}
+
+
